+++ spec/app/models/building_address.rb
@@ -1,0 +1,7 @@
+class BuildingAddress
+  include Mongoid::Document
+  attr_accessible
+  attr_accessible :city, as: :admin
+  embedded_in :building
+  field :city, type: String
+end